+++ src/main/java/amaralus/apps/hackandslash/gameplay/FontData.java
@@ -1,7 +1,7 @@
 package amaralus.apps.hackandslash.gameplay;
 
-import amaralus.apps.hackandslash.graphics.entities.gpu.Texture;
 import amaralus.apps.hackandslash.graphics.entities.gpu.buffer.VertexArraysObject;
+import amaralus.apps.hackandslash.graphics.entities.gpu.texture.Texture;
 import org.lwjgl.stb.STBTTBakedChar;
 import org.lwjgl.stb.STBTTFontinfo;
