+++ src/main/java/amaralus/apps/hackandslash/gameplay/GameplayManager.java
@@ -8,9 +8,9 @@ import amaralus.apps.hackandslash.gameplay.state.action.InputEventProcessingActi
 import amaralus.apps.hackandslash.gameplay.state.action.MessageProcessingStateAction;
 import amaralus.apps.hackandslash.graphics.RendererService;
 import amaralus.apps.hackandslash.graphics.Window;
-import amaralus.apps.hackandslash.graphics.entities.gpu.Texture;
 import amaralus.apps.hackandslash.graphics.entities.gpu.buffer.IntVertexBufferObject;
-import amaralus.apps.hackandslash.graphics.entities.gpu.factory.TextureFactory;
+import amaralus.apps.hackandslash.graphics.entities.gpu.texture.Texture;
+import amaralus.apps.hackandslash.graphics.entities.gpu.texture.TextureFactory;
 import amaralus.apps.hackandslash.graphics.entities.sprites.Sprite;
 import amaralus.apps.hackandslash.graphics.entities.sprites.SpriteRenderComponent;
 import amaralus.apps.hackandslash.io.FileLoadService;
@@ -34,14 +34,14 @@ import static amaralus.apps.hackandslash.gameplay.entity.EntityStatus.REMOVE;
 import static amaralus.apps.hackandslash.gameplay.entity.EntityStatus.SLEEPING;
 import static amaralus.apps.hackandslash.graphics.entities.Color.CYAN;
 import static amaralus.apps.hackandslash.graphics.entities.Color.WHITE;
-import static amaralus.apps.hackandslash.graphics.entities.gpu.Texture.Filter.CLAMP_TO_EDGE;
-import static amaralus.apps.hackandslash.graphics.entities.gpu.Texture.Filter.LINEAR;
-import static amaralus.apps.hackandslash.graphics.entities.gpu.Texture.ParameterName.*;
-import static amaralus.apps.hackandslash.graphics.entities.gpu.Texture.PixelFormat.RED;
 import static amaralus.apps.hackandslash.graphics.entities.gpu.buffer.BufferType.ARRAY_BUFFER;
 import static amaralus.apps.hackandslash.graphics.entities.gpu.buffer.BufferUsage.DYNAMIC_DRAW;
 import static amaralus.apps.hackandslash.graphics.entities.gpu.buffer.factory.VaoFactory.newVao;
 import static amaralus.apps.hackandslash.graphics.entities.gpu.buffer.factory.VboFactory.floatBuffer;
+import static amaralus.apps.hackandslash.graphics.entities.gpu.texture.PixelFormat.RED;
+import static amaralus.apps.hackandslash.graphics.entities.gpu.texture.TextureFilter.CLAMP_TO_EDGE;
+import static amaralus.apps.hackandslash.graphics.entities.gpu.texture.TextureFilter.LINEAR;
+import static amaralus.apps.hackandslash.graphics.entities.gpu.texture.TextureParameterName.*;
 import static amaralus.apps.hackandslash.graphics.scene.NodeRemovingStrategy.CASCADE;
 import static amaralus.apps.hackandslash.io.events.keyboard.KeyCode.*;
 import static amaralus.apps.hackandslash.io.events.mouse.MouseButton.MOUSE_BUTTON_LEFT;
