+++ src/main/java/amaralus/apps/hackandslash/graphics/entities/sprites/FramesStrip.java
@@ -1,6 +1,6 @@
 package amaralus.apps.hackandslash.graphics.entities.sprites;
 
-import amaralus.apps.hackandslash.graphics.entities.gpu.Texture;
+import amaralus.apps.hackandslash.graphics.entities.gpu.texture.Texture;
 import amaralus.apps.hackandslash.io.data.FrameStripData;
 import org.joml.Vector2f;
